+++ ros/src/can_com/scripts/start_vcan.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+sudo modprobe vcan
+sudo ip link add dev vcan0 type vcan
+sudo ip link set vcan0 up
